+++ tests/test_file_namers.py
@@ -1,5 +1,6 @@
 import pytest
 import xarray as xr
+from roocs_utils.exceptions import InvalidProject
 from roocs_utils.parameter.param_utils import time_interval
 
 from _common import C3S_CORDEX_NAM_PR, CMIP5_TAS, CMIP6_SICONC
@@ -64,7 +65,7 @@ def test_StandardFileNamer_no_project_match():
     mock_ds = Thing()
     mock_ds.attrs = {}
 
-    with pytest.raises(KeyError):
+    with pytest.raises(InvalidProject):
         s.get_file_name(mock_ds)